+++ Back-End/OhEasy-Data-API/src/main/java/kr/or/oheasy/hrm/dao/AuthDao.java
@@ -1,4 +1,4 @@
-package kr.or.oheasy.dao;
+package kr.or.oheasy.hrm.dao;
 
 import kr.or.oheasy.vo.LoginVO;
