+++ Back-End/OhEasy-Data-API/src/main/java/kr/or/oheasy/hrm/service/AuthService.java
@@ -1,10 +1,10 @@
-package kr.or.oheasy.service;
+package kr.or.oheasy.hrm.service;
 
 import org.apache.ibatis.session.SqlSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import kr.or.oheasy.dao.AuthDao;
+import kr.or.oheasy.hrm.dao.AuthDao;
 import kr.or.oheasy.vo.LoginVO;
 
 @Service
