+++ Back-End/OhEasy-Data-API/src/main/java/kr/or/oheasy/hrm/service/HrService.java
@@ -1,4 +1,4 @@
-package kr.or.oheasy.service;
+package kr.or.oheasy.hrm.service;
 
 import java.util.List;
 
@@ -6,7 +6,7 @@ import org.apache.ibatis.session.SqlSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import kr.or.oheasy.dao.HrDao;
+import kr.or.oheasy.hrm.dao.HrDao;
 import kr.or.oheasy.vo.HrEmpMstVO;
 
 @Service
