+++ Back-End/OhEasy-Data-API/src/main/resources/mapper/HrMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="kr.or.oheasy.dao.HrDao">
+<mapper namespace="kr.or.oheasy.hrm.dao.HrDao">
 
 <select id="getAllEmpList" resultType="kr.or.oheasy.vo.HrEmpMstVO">
 	select * from hr_emp_mst
